+++ Breakout/Ball.h
@@ -5,7 +5,7 @@ class Ball : public GameObject
 {
 public:
 	const static int m_maxSpeed = 15;
-	const int m_startingAngle = 270;
+	const float m_startingAngle = 270;
 	const int m_startingSpeed = 6;
 	const int m_radius = 25;
 public:
@@ -13,7 +13,7 @@ public:
 	Ball();
 
 	void SetSpeed(int speed);
-	void SetAngle(int angle);
+	void SetAngle(float angle);
 private:
 	void Update() override;
 };
