+++ Breakout/Text.cpp
@@ -10,8 +10,8 @@ Text::Text(std::string text, Vector2 size, const SDL_Color& color, Vector2 offse
 
 void Text::Render()
 {
-	m_rect.x = GetPosition().x;
-	m_rect.y = GetPosition().y;
+	m_rect.x = static_cast<int>(GetPosition().x);
+	m_rect.y = static_cast<int>(GetPosition().y);
 	SDL_RenderCopy(TextureManager::GetManager()->GetRenderer(), m_texture, NULL, &m_rect.SDLRect());
 }
 
@@ -28,8 +28,8 @@ Vector2 Text::GetPosition()
 
 void Text::SetSize(Vector2 size)
 {
-	m_rect.w = size.x;
-	m_rect.h = size.y;
+	m_rect.w = static_cast<int>(size.x);
+	m_rect.h = static_cast<int>(size.y);
 }
 void Text::SetOffset(Vector2 offset)
 {
